pyproject.toml
[project]
name = "dj-commerce-api"
version = "1.0.0"
description = "Fully featured E-Commerce API using Django and Django Rest Framework"
authors = [
    { name = "Uzzal Hossain", email = "de3dm3nu11@gmail.com" }
]
dependencies = ["django~=4.2.1", "django-environ~=0.10.0", "psycopg[binary]~=3.1.9", "djangorestframework~=3.14.0", "dj-rest-auth~=4.0.0", "django-allauth~=0.54.0", "drf-spectacular~=0.26.2", "djangorestframework-simplejwt~=5.2.2", "celery~=5.2.7", "redis~=4.5.5", "django-celery-beat~=2.5.0", "django-celery-results~=2.5.1", "django-celery-email~=3.0.0", "django-smart-selects~=1.6.0", "django-phonenumber-field[phonenumbers]~=7.1.0", "django-versatileimagefield~=3.0", "django-mptt~=0.14.0", "django-lifecycle~=1.0.0"]
readme = "README.md"
requires-python = ">= 3.11"

[project.optional-dependencies]
tests = ["pytest~=7.3.1", "pytest-django~=4.5.2", "pytest-sugar~=0.9.7", "factory-boy~=3.2.1", "pytest-icdiff~=0.6", "pytest-clarity~=1.0.1", "pytest-html~=3.2.0", "coverage~=7.2.5", "pytest-factoryboy~=2.5.1", "pytest-cov~=4.0.0", "freezegun~=1.2.2", "faker-e164~=0.1.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = ["reorder-python-imports~=3.9.0", "setuptools~=67.7.2", "flower~=1.2.0", "django-extensions~=3.2.1", "pywatchman~=1.4.1", "ipython~=8.13.2"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "--color=yes --html=reports/test_report.html"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning",]

[tool.coverage]
# fail_under = 80
[tool.coverage.run]
branch = true
source = [".",]

